# IGVF-2864 Group and sort sequence file tables