--- a/package.json
+++ b/package.json
@@ -67,6 +67,7 @@
     "@types/cypress": "^1.1.3",
     "@types/jest": "^30.0.0",
     "@types/lodash": "^4.14.195",
+    "@types/xxhashjs": "^0.2.4",
     "auth0-js": "^9.19.0",
     "coveralls": "^3.1.1",
     "cypress": "^14.5.1",