… components/sequencing-file-table.js
… lib/__tests__/files.test.ts
… lib/files.ts
… pages/cell-models/[view].tsx
… pages/tissue-summary/[taxa].tsx